update according review comments
